Move gap filling to a new module gap_filling.rs.
